requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.6.1
gunicorn==23.0.0
h11==0.14.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
whitenoise==6.7.0
